wip push to collaborate with Johannes

+++ dataland-e2etests/src/test/kotlin/org/dataland/e2etests/tests/DataOwnerControllerTest.kt
@@ -60,33 +60,33 @@ class DataOwnerControllerTest {
             .getTData(1)[0]
 
         assertFailingApiUploadToCompany(firstCompanyId, frameworkSampleData, false)
-        assertFailingApiUploadToCompany(secondCompanyId, frameworkSampleData, false)
+        /*    assertFailingApiUploadToCompany(secondCompanyId, frameworkSampleData, false)
 
-        jwtHelper.authenticateApiCallsWithJwtForTechnicalUser(TechnicalUser.Admin)
-        val dataOwnersAfterPostRequest = apiAccessor.companyDataControllerApi.postDataOwner(
-            firstCompanyId,
-            dataReaderUserId,
-        )
-        validateDataOwnersForCompany(firstCompanyId, listOf(dataReaderUserId), dataOwnersAfterPostRequest)
-        checkDuplicateRequest(firstCompanyId)
+            jwtHelper.authenticateApiCallsWithJwtForTechnicalUser(TechnicalUser.Admin)
+            val dataOwnersAfterPostRequest = apiAccessor.companyDataControllerApi.postDataOwner(
+                firstCompanyId,
+                dataReaderUserId,
+            )
+            validateDataOwnersForCompany(firstCompanyId, listOf(dataReaderUserId), dataOwnersAfterPostRequest)
+            assertThatRepostingSameDataOwnerHasNoEffect(firstCompanyId)
 
-        jwtHelper.authenticateApiCallsWithJwtForTechnicalUser(TechnicalUser.Reader)
+            jwtHelper.authenticateApiCallsWithJwtForTechnicalUser(TechnicalUser.Reader)
 
-        assertSucceedingApiUploadToCompany(firstCompanyId, frameworkSampleData)
-        assertFailingApiUploadToCompany(secondCompanyId, frameworkSampleData, false)
+            assertSucceedingApiUploadToCompany(firstCompanyId, frameworkSampleData)
+            assertFailingApiUploadToCompany(secondCompanyId, frameworkSampleData, false)
 
-        jwtHelper.authenticateApiCallsWithJwtForTechnicalUser(TechnicalUser.Admin)
-        val dataOwnersAfterRemovingUser =
-            apiAccessor.companyDataControllerApi.deleteDataOwner(firstCompanyId, dataReaderUserId)
-        jwtHelper.authenticateApiCallsWithJwtForTechnicalUser(TechnicalUser.Reader)
+            jwtHelper.authenticateApiCallsWithJwtForTechnicalUser(TechnicalUser.Admin)
+            val dataOwnersAfterRemovingUser =
+                apiAccessor.companyDataControllerApi.deleteDataOwner(firstCompanyId, dataReaderUserId)
+            jwtHelper.authenticateApiCallsWithJwtForTechnicalUser(TechnicalUser.Reader)
 
-        assertFailingApiUploadToCompany(firstCompanyId, frameworkSampleData, false)
-        assertFailingApiUploadToCompany(secondCompanyId, frameworkSampleData, false)
+            assertFailingApiUploadToCompany(firstCompanyId, frameworkSampleData, false)
+            assertFailingApiUploadToCompany(secondCompanyId, frameworkSampleData, false)
 
-        assertEquals(dataOwnersAfterRemovingUser, CompanyDataOwners(firstCompanyId.toString(), mutableListOf()))
+            assertEquals(dataOwnersAfterRemovingUser, CompanyDataOwners(firstCompanyId.toString(), mutableListOf())) */
     }
 
-    private fun checkDuplicateRequest(companyId: UUID) {
+    private fun assertThatRepostingSameDataOwnerHasNoEffect(companyId: UUID) {
         val dataOwnersAfterDuplicatePostRequest = apiAccessor.companyDataControllerApi
             .postDataOwner(companyId, dataReaderUserId)
         validateDataOwnersForCompany(companyId, listOf(dataReaderUserId), dataOwnersAfterDuplicatePostRequest)
